Fix name of temporal directory
By using `arguments` it was being always set the temporal directory name to `/tmp/_tmpNameCreated` because it caught the name of the function. After removing it the folders are being created with a random name as it should be. Another maybe better option would be to use `Array.slice()` or `Array.shift()` to remove the `fn` function so options could be passed to underlying `tmp` module, like `unsafeCleanup`. What do you think?